% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recur.R
\docType{methods}
\name{as.character,Recur-method}
\alias{as.character,Recur-method}
\title{Convert An Recur Object to A Character Vector}
\usage{
\S4method{as.character}{Recur}(x, ...)
}
\arguments{
\item{x}{An Recur object.}

\item{...}{Other arguments for future usage.}
}
\description{
Summarize and convert the recurrent episodes for each subjects into
character strings.
}
\details{
This function is intended to be a helper function for the `show()` method of
`Recur` objects.  To be precise, the function set the maximum number of
recurrent episodes for each subject to be `max(2L,
as.integer(getOption("reda.Recur.maxPrint")))`.  By default, at most three
recurrent episodes will be summarized for printing.  When subjects having
more than three recurrent episodes, the first
`getOption("reda.Recur.maxPrint") - 1` number of recurrent episodes and the
last one will be summarized.  One may use `options()` to adjust the setting.
For example, the default value is equivalent to `options(reda.Recur.maxPrint
= 3)`.
}
