% Please edit documentation in R/baseline.R
\docType{methods}
\name{baseline,heart-method}
\alias{baseline,heart-method}
\title{Estimated Baseline Rate Function.}
\usage{
baseline(object, ...)
}
\arguments{
\item{object}{heart object.}

\item{...}{further arguments.}
}
\value{
A named vector.
}
\description{
Extract estiamted baseline rate function from heart object.
}
