% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{baseRate.rateReg-class}
\alias{baseRate.rateReg-class}
\title{An S4 Class Representing Estimated Baseline Rate Function}
\description{
An S4 class that represents the estimated baseline rate function from model.
The function \code{\link{baseRate}} produces objects of this class.
}
\section{Slots}{

\describe{
\item{\code{baseRate}}{A data frame.}

\item{\code{level}}{A numeric value.}
}}

\seealso{
\code{\link{baseRate,rateReg-method}}
}
