% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{baseRateReg-class}
\alias{baseRateReg-class}
\title{An S4 Class to Represent Estimated Baseline Rate Function}
\description{
An S4 class that represents Estimated Baseline Rate Function from model.
\code{\link{baseRate}} produces objects of this class.
}
\section{Slots}{

\describe{
\item{\code{baseRate}}{A data frame.}

\item{\code{level}}{A numeric value.}
}}
\seealso{
\code{\link{baseRate,rateReg-method}} for details of slots.
}

