% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recur.R
\name{check_Recur}
\alias{check_Recur}
\title{Checks for Recurrent Event Data}
\usage{
check_Recur(x, check = c("hard", "soft", "none"))
}
\arguments{
\item{x}{An \code{Recur} object.}

\item{check}{A character value specifying how to perform the checks for
recurrent event data.  Errors or warnings will be thrown, respectively,
if the \code{check} is specified to be \code{"hard"} (by default) or
\code{"soft"}.  If \code{check = "none"} is specified, no data checking
procedure will be run.}
}
\value{
An \code{Recur} object invisibly.
}
\description{
Perform several checks for recurrent event data and update object
attributions if some rows of the contained data (in the \code{.Data} slot)
have been removed by such as \code{na.action}.
}
