% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\docType{methods}
\name{coef,rateReg-method}
\alias{coef,rateReg-method}
\title{Estimated Coefficients of Covariates}
\usage{
\S4method{coef}{rateReg}(object, ...)
}
\arguments{
\item{object}{A \code{rateReg} object.}

\item{...}{Other arguments for future usage.}
}
\value{
A named numeric vector.
}
\description{
\code{coef,rateReg-method} is an S4 class method that extracts estimated
coefficients of covariates from \code{rateReg} object produced by function
\code{\link{rateReg}}.
}
\examples{
## See examples given in function rateReg.
}
\seealso{
\code{\link{rateReg}} for model fitting;
\code{\link{confint,rateReg-method}} for confidence intervals
for covariate coefficients;
\code{\link{summary,rateReg-method}} for summary of a fitted model.
}
