\name{heart-imports}
\alias{summary}
\alias{.__T__[<-:base}
\alias{.__T__[:base}
\alias{.__T__[[<-:base}
\alias{.__T__$<-:base}
\alias{.__T__$:base}
\alias{show}
\alias{coef}
\alias{confint}
\docType{import}
\title{Objects imported from other packages}
\description{
These objects are imported from other packages. Follow the links to their documentation.
\describe{
  \item{base}{\code{\link[base:summary]{summary}}}
  \item{ggplot2}{\code{\link[ggplot2:.__T__[<-:base]{.__T__[<-:base}}, \code{\link[ggplot2:.__T__[:base]{.__T__[:base}}, \code{\link[ggplot2:.__T__[[<-:base]{.__T__[[<-:base}}, \code{\link[ggplot2:.__T__$<-:base]{.__T__$<-:base}}, \code{\link[ggplot2:.__T__$:base]{.__T__$:base}}}
  \item{methods}{\code{\link[methods:show]{show}}}
  \item{stats}{\code{\link[stats:coef]{coef}}, \code{\link[stats:confint]{confint}}}
}}
