% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recur.R
\name{is.Recur}
\alias{is.Recur}
\title{Is the xect from the Recur class?}
\usage{
is.Recur(x)
}
\arguments{
\item{x}{An \code{R} xect.}
}
\value{
A logical value.
}
\description{
Return \code{TRUE} if the specified xect is from the \code{\link{Recur}}
class, \code{FALSE} otherwise.
}
