% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{mcf.formula-class}
\alias{mcf.formula-class}
\title{An S4 Class Representing Sample MCF}
\description{
An S4 class that represents sample mean cumulative function (MCF) from data.
The function \code{\link{mcf}} produces objects of this class.
}
\section{Slots}{

\describe{
\item{\code{formula}}{Formula.}

\item{\code{data}}{A data frame.}

\item{\code{MCF}}{A data frame.}

\item{\code{origin}}{A named numeric vector.}

\item{\code{multiGroup}}{A logical value.}

\item{\code{variance}}{A character vector.}

\item{\code{logConfInt}}{A logical value.}

\item{\code{level}}{A numeric value.}
}}

\seealso{
\code{\link{mcf,formula-method}}.
}
