% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{mcf.rateReg-class}
\alias{mcf.rateReg-class}
\title{An S4 Class Respresenting Estimated MCF from a Fitted Model}
\description{
An S4 class that represents estimated mean cumulative function (MCF) from
Models. The function \code{\link{mcf}} produces objects of this class.
}
\section{Slots}{

\describe{
\item{\code{call}}{Function call.}

\item{\code{formula}}{Formula.}

\item{\code{spline}}{A character.}

\item{\code{knots}}{A numeric vector.}

\item{\code{degree}}{A nonnegative integer.}

\item{\code{Boundary.knots}}{A numeric vector.}

\item{\code{newdata}}{A numeric matrix.}

\item{\code{MCF}}{A data frame.}

\item{\code{level}}{A numeric value between 0 and 1.}

\item{\code{na.action}}{A length-one character vector.}

\item{\code{control}}{A list.}

\item{\code{multiGroup}}{A logical value.}
}}

\seealso{
\code{\link{mcf,rateReg-method}}
}
