% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{mcfDiff-class}
\alias{mcfDiff-class}
\alias{sampleMcfDiff-class}
\title{An S4 Class Representing Sample MCF Difference}
\description{
An S4 class that represents the difference between two sample mean
cumulative functions from data.  The function \code{\link{mcfDiff}}
produces objects of this class.
}
\section{Slots}{

\describe{
\item{\code{MCF}}{A data frame.}

\item{\code{origin}}{A named numeric vector.}

\item{\code{variance}}{A character vector.}

\item{\code{logConfInt}}{A logical value.}

\item{\code{level}}{A numeric value.}
}}

\seealso{
\code{\link{mcfDiff}} for details of slots.
}
