% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcfDiff.R
\name{mcfDiff}
\alias{mcfDiff}
\title{Estimate Difference Between Two MCF curves}
\usage{
mcfDiff(mcf1, mcf2 = NULL, level = 0.95, ...)
}
\arguments{
\item{mcf1}{A \code{sampleMcf} object representing the MCF for one or two
groups.}

\item{mcf2}{An optional second \code{sampleMcf} object.}

\item{level}{An optional numeric value indicating the confidence level
required. The default value is 0.95.}

\item{...}{Other arguments for future usage.}
}
\value{
A \code{mcfDiff} object that contains the following slots:
\itemize{
    \item \code{MCF}: Estimated Mean cumulative function Difference
        at each time point.
    \item \code{origin}: Time origins of the two groups.
    \item \code{variance}: The method used for variance estimates.
    \item \code{logConfInt}: A logical value indicating whether normality
        is assumed for \code{log(MCF)} instead of MCF itself.
        For \code{mcfDiff} object, it is always \code{FALSE}.
    \item \code{level}: Confidence level specified.
}
}
\description{
This function estimates MCF difference between two groups. Both point
estimates and confidence intervals are returned.
}
\examples{
## See examples given for function mcf.
}
\references{
Nelson, W. B. (2003). \emph{Recurrent events data analysis for product
repairs, disease recurrences, and other applications}. Society for
Industrial and Applied Mathematics.

Doganaksoy, N., & Nelson, W. (1998). A Method to Compare Two Samples of
Recurrence Data. \emph{Lifetime Data Analysis}, 4(1), 51--63.
}
