% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{mcfDiff.test-class}
\alias{mcfDiff.test-class}
\title{An S4 Class Representing the Two-Sample Pseudo-Score Test Results}
\description{
An S4 class that represents the results of the two-sample pseudo-score tests
between two sample mean cumulative functions.  The function
\code{\link{mcfDiff.test}} produces objects of this class.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A numeric matrix.}

\item{\code{testVariance}}{A character vector.}
}}

\seealso{
\code{\link{mcfDiff.test}}
}
