% Please edit documentation in R/plot.R
\name{plot.baseline}
\alias{plot.baseline}
\title{plot baseline.}
\usage{
\method{plot}{baseline}(object, CI = TRUE, level = 0.95)
}
\description{
\code{plot} produces the plot of baseline function.
}
\details{
This is a function using R base plotting system,
which probably can be rewritten using ggplot2 later.
(This is a test Roxygen comments)
}
