% Please edit documentation in R/plot.R
\name{plot_MCF}
\alias{plot_MCF}
\title{Plot Mean Cumulative Function (MCF) of Baseline Rate Function.}
\usage{
plot_MCF(object, CI = TRUE, level = 0.95)
}
\description{
\code{plot_MCF} produces the plot of baseline function.
}
\details{
This is a function using R base plotting system,
which probably can be rewritten using ggplot2 later.
(This is a test Roxygen comments)
}
