% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{Recur-class}
\alias{Recur-class}
\title{An S4 Class Representing Formula Response for Recurrent Event Data}
\description{
The class \code{Recur} is an S4 that represents a formula response for
recurrent event data model.  The function \code{\link{Recur}} produces
objects of this class.  See ``Slots'' for details.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A numeric matrix that consists of the following columns:
\itemize{ \item \code{time1}: the beginning of time segements; \item
\code{time2}: the end of time segements; \item \code{id}: Identificators
of subjects; \item \code{event}: Event indicators; \item:
\code{terminal}: Indicators of terminal events.}}

\item{\code{ID}}{A charactrer vector for original identificators of subjects.}

\item{\code{ord}}{An integer vector for increasingly ordering data by \code{id},
\code{time2}, and \code{- event}.}

\item{\code{rev_ord}}{An integer vector for reverting the ordering of the sorted
data to its original ordering.}

\item{\code{first_idx}}{An integer vector indicating the first record of each
subject in the sorted data.}

\item{\code{last_idx}}{An integer vector indicating the last record of each subject
in the sorted data.}

\item{\code{check}}{A character string indicating how the data checking is
performed.}
}}

\seealso{
\code{\link{Recur}}
}
