% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recur.R
\name{Recur}
\alias{Recur}
\title{Formula Response for Recurrent Event Data}
\usage{
Recur(time, id, event, death, origin, check = TRUE, ...)
}
\arguments{
\item{time}{A numerical vector representing the time of reccurence event or
censoring, or a list with elements named \code{"time1"} and
\code{"time2"} for specifying the follow-up of recurrent events.  In the
latter case, function \code{\%to\%} (or \code{\%2\%}) can be used for
ease of typing.  In addition to numeric values, \code{Date} and
\code{difftime} are allowed and converted to numeric values.  An error
will be thrown if this argument is not specified.}

\item{id}{Subject identificators. It can be numeric vector, character
vector, or a factor vector.  If it is left unspecified, \code{Recur}
will assume that each row represents a subject.}

\item{event}{A numeric vector that represents event indicators, event costs,
or event types. Logical vector is allowed and converted to numeric
vector. Censoring is indicated by non-positive values.}

\item{death}{A numeric vector}

\item{origin}{The time origin of each subject or process. In addition to
numeric values, \code{Date} and \code{difftime} are also supported and
converted to numeric values.  Different subjects may have different
origins. However, one subject must have the same origin.}

\item{check}{A logical value suggesting whether to perform data checking
procedure. The default value is \code{TRUE}. \code{FALSE} should be set
with caution and only for processed data already in recerruent event
data framework.}

\item{...}{Other arguments for future usage.  A warning will be thrown if
any invalid argument is specified.}
}
\value{
An \code{Recur} object.
}
\description{
Create an S4 class object that represents formula response for recurrent
event data with optional checking procedures embedded.
}
\details{
This is a successor function of the deprecated function \code{Survr}.
The checking rules apply to each subject respectively and include that
\itemize{
    \item Subject identification, event times, censoring time, and event
        indicator cannot be missing or contain missing values.
    \item There has to be only one censoring time not earlier than
        any event time.
    \item The time origin has to be the same and not later than any event
        time.
}
}
