% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reda.R
\docType{package}
\name{reda-package}
\alias{reda-package}
\title{Recurrent Event Data Analysis}
\description{
The R package \pkg{reda} provides functions for simulating, exploring and
modeling recurrent event data.
}
\details{
The main functions are summarized as follows:
\itemize{

\item \code{simEventData}: Simulating survival, recurrent event, or multiple
event data from stochastic process point of view.

\item \code{mcf}: Mean cumulative function (MCF) estimates from a fitted
parametric gamma frailty model or the nonparametic MCF estimator (the
Nelson-Aelen estimator of the cumulative hazard function) from the sample
data.

\item \code{mcfDiff}: Comparing two-sample MCFs by the pseudo-score tests
and estimating their difference over time.

\item \code{rateReg}: Fitting parametric Gamma fraitly model with spline
baseline rate function.
}

See the package vignettes for introduction and demonstration.
}
