% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{sampleMcf-class}
\alias{sampleMcf-class}
\title{An S4 Class to Represent Sample MCF}
\description{
An S4 class that represents sample mean cumulative function (MCF) from data.
\code{\link{mcf}} produces objects of this class.
}
\section{Slots}{

\describe{
\item{\code{formula}}{Formula.}

\item{\code{MCF}}{A data frame.}

\item{\code{multiGroup}}{A logical value.}

\item{\code{na.action}}{A length-one character vector.}

\item{\code{variance}}{A character.}

\item{\code{logConfInt}}{A logical value.}

\item{\code{level}}{A numeric value.}
}}
\seealso{
\code{\link{mcf,formula-method}} for details of slots.
}

