% Please edit documentation in R/show_heart.R
\docType{methods}
\name{show,heart-method}
\alias{show,heart-method}
\title{\code{show} prints heart objects.}
\usage{
\S4method{show}{heart}(object)
}
\arguments{
\item{object}{object from class \code{heart}}
}
\description{
S4 class function to print \code{heart} objects
generated from function \code{heart}.
}
\details{
The contents printed by this function \code{show} is
a brief summary of the \code{heart} object.
}
\seealso{
\code{\link[heart]{heart}}
}
