% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\docType{methods}
\name{show-method}
\alias{show-method}
\alias{show,rateReg-method}
\alias{show,summary.rateReg-method}
\alias{show,summaryRateReg-method}
\alias{show,mcf.formula-method}
\alias{show,mcf.rateReg-method}
\alias{show,simEvent-method}
\alias{show,mcfDiff-method}
\alias{show,mcfDiff.test-method}
\title{Show an object.}
\usage{
\S4method{show}{rateReg}(object)

\S4method{show}{summary.rateReg}(object)

\S4method{show}{mcf.formula}(object)

\S4method{show}{mcf.rateReg}(object)

\S4method{show}{simEvent}(object)

\S4method{show}{mcfDiff}(object)

\S4method{show}{mcfDiff.test}(object)
}
\arguments{
\item{object}{An object used to dispatch a method.}
}
\description{
An S4 class generic function that displays certain object.
}
\details{
\itemize{
  \item For \code{rateReg} object,
      it prints out brief summary of the fitted model.
  \item For \code{summary.rateReg} object,
      it prints out summary of the fitted model.
  \item For \code{mcf.sample} object,
      it prints out the function call, formula and
      the sample MCF data frame.
  \item For \code{mcf.rateReg} object,
      it prints formula, new data, confidence level,
      and the estimated MCF data frame.
}
}
\seealso{
\code{\link{rateReg}} for model fitting;
\code{\link{summary,rateReg-method}} for summary of a fitted model;
\code{\link{mcf}} for estimation of MCF.
}
