% Please edit documentation in R/show.R
\docType{methods}
\name{show,summary.heart-method}
\alias{show,summary.heart-method}
\title{\code{show} prints summary.heart objects.}
\usage{
\S4method{show}{summary.heart}(object)
}
\arguments{
\item{object}{object from class \code{summary.heart}}
}
\description{
S4 class function to print \code{summary.heart} objects
generated from function \code{summary}.
}
\details{
The contents printed by this function \code{show} is
a summary of the \code{summary.heart} object.
}
\seealso{
\code{\link[heart]{heart}}, \code{\link[heart]{summary}},
}
