% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{simRec-class}
\alias{simRec-class}
\title{An S4 Class for Simulated Recurrent Event or Survival Times}
\description{
An S4 class that represents simulated recurrent event or survival time from
one stochastic process. The function \code{\link{simRec}} produces objects
of this class.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A numerical vector of possibly length zero.}

\item{\code{call}}{A function call.}

\item{\code{z}}{A list.}

\item{\code{zCoef}}{A list.}

\item{\code{rho}}{A list.}

\item{\code{rhoCoef}}{A numerical vector.}

\item{\code{frailty}}{A list.}

\item{\code{origin}}{A numeric vector.}

\item{\code{endTime}}{A numeric vector.}

\item{\code{censoring}}{A list.}

\item{\code{recurrent}}{A logical vector.}

\item{\code{method}}{A character vector.}
}}

\seealso{
\code{\link{simRec}} for details of slots.
}
