% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simuDat}
\alias{simuDat}
\title{Simulated Sample Dataset for Demonstration}
\format{A data frame with 500 rows and 5 variables.}
\description{
A simulated data frame with covariates named
\code{ID}, \code{time}, \code{event}, \code{group} and \code{x1},
where
\itemize{
    \item \code{ID}: Subjects identification;
    \item \code{time}: Event or censoring time;
    \item \code{event}: Event indicator, 1 = event, 0 = censored;
    \item \code{group}: Treatment group indicator;
    \item \code{x1}: Continuous variable.
}
}
\details{
The sample dataset is originally simulated by the thinning
method developed by Lewis and Shedler (1979) and
further processed for a better demonstration purpose.
See Fu et al. (2014) for details also.
}
\references{
Lewis, P. A., & Shedler, G. S. (1979).
Simulation of nonhomogeneous Poisson processes by thinning.
\emph{Naval Research Logistics Quarterly}, 26(3), 403--413.

Fu, H., Luo, L., & Qu Y. (2014). Hypoglycemic Events Analysis via
Recurrent Time-to-Event (HEART) Models.
\emph{Journal of biopharmaceutical statistics}, Epub 2014 Dec 1.
}

