% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,Recur-method}
\alias{summary,Recur-method}
\title{Summarize an \code{Recur} object}
\usage{
\S4method{summary}{Recur}(object, ...)
}
\arguments{
\item{object}{An \code{Recur} object.}

\item{...}{Other arguments not used.}
}
\value{
\code{summary.Recur} object.
}
\description{
Summarize an \code{Recur} object
}
