% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{summary.rateReg-class}
\alias{summary.rateReg-class}
\title{An S4 Class Representing Summary of a Fitted Model}
\description{
The class \code{summary.rateReg} is an S4 class with selective slots of
\code{rateReg} object.  See ``Slots'' for details.  The function
\code{\link{summary,rateReg-method}} produces objects of this class.
}
\section{Slots}{

\describe{
\item{\code{call}}{Function call.}

\item{\code{spline}}{A character.}

\item{\code{knots}}{A numeric vector.}

\item{\code{Boundary.knots}}{A numeric vector.}

\item{\code{covarCoef}}{A numeric matrix.}

\item{\code{frailtyPar}}{A numeric matrix.}

\item{\code{degree}}{A nonnegative integer.}

\item{\code{baseRateCoef}}{A numeric matrix.}

\item{\code{logL}}{A numeric value.}
}}

\seealso{
\code{\link{summary,rateReg-method}}
}
