% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{summary.Recur-class}
\alias{summary.Recur-class}
\title{An S4 Class for Summarized Recur Object}
\description{
An S4 Class for Summarized Recur Object
}
\section{Slots}{

\describe{
\item{\code{call}}{A function call.}

\item{\code{sampleSize}}{An integer representing the sample size (number of
subjects).}

\item{\code{reSize}}{An integer representing the number of recurrent events.}

\item{\code{avgReSize}}{A numeric value representing the average number of recurrent
events per subject.}

\item{\code{propTem}}{A numeric value representing the proportion of subjects having
terminal event.}

\item{\code{medFU}}{A numeric value for median follow-up time.}

\item{\code{medTem}}{A numeric value for median survival time of the terminal
events.}
}}

