% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valveSeats}
\alias{valveSeats}
\title{Valve Seats Dataset}
\format{A data frame with 89 rows and 3 variables.}
\description{
Valve seats wear out in certain diesel engines, each with 16 valve seats.
The dataset served as an example of recurrence data in Nelson (1995),
which consists of valve-seat replacements on 41 engines in a fleet.
The covariates are named
\code{ID}, \code{Days}, and \code{No.}, where
\itemize{
    \item \code{ID}: The engine number;
    \item \code{Days}: Engine age in days;
    \item \code{No.}: Event indicator, '1' for a valve-seat replacement
        and, '0' for the censoring age of an engine.
}
}
\references{
Nelson, W. (1995), Confidence Limits for Recurrence
Data-Applied to Cost or Number of Product Repairs, \emph{Technometrics},
37, 147--157.
}
