% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\docType{methods}
\name{baseRate}
\alias{baseRate}
\alias{baseRate,rateReg-method}
\title{Estimated Coefficients of Baseline Rate Function}
\usage{
baseRate(object, ...)

\S4method{baseRate}{rateReg}(object, ...)
}
\arguments{
\item{object}{An object used to dispatch a method.}

\item{...}{Other arguments for future usage.}
}
\value{
A named numeric vector.
}
\description{
An S4 class generic function that returns the estimated coefficients
of baseline rate function. For \code{\link{rateReg-class}} object,
it returns either coefficients of pieceswise (including one piece)
constant rate function or coefficients of B-spline bases.
}
\section{Methods (by class)}{
\itemize{
\item \code{rateReg}: Extract estiamted coefficients of
baseline rate function from \code{rateReg-class} object.
}}
\examples{
## See examples given in function rateReg.
}
\seealso{
\code{\link{rateReg}} for model fitting;
\code{\link{summary,rateReg-method}} for summary of a fitted model.
}

