\name{heart-imports}
\alias{summary}
\alias{show}
\alias{coef}
\alias{confint}
\docType{import}
\title{Objects imported from other packages}
\description{
These objects are imported from other packages. Follow the links to their documentation.
\describe{
  \item{base}{\code{\link[base:summary]{summary}}}
  \item{methods}{\code{\link[methods:show]{show}}}
  \item{stats}{\code{\link[stats:coef]{coef}}, \code{\link[stats:confint]{confint}}}
}}
