% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcf-generic.R, R/mcf-formula.R, R/mcf-rateReg.R
\docType{methods}
\name{mcf}
\alias{mcf}
\alias{mcf,formula-method}
\alias{mcf,rateReg-method}
\title{Mean Cumulative Function (MCF)}
\usage{
mcf(object, ...)

\S4method{mcf}{formula}(object, data, subset, na.action,
  variance = c("LawlessNadeau", "Poisson", "bootstrap"), logConfInt = FALSE,
  level = 0.95, control = list(), ...)

\S4method{mcf}{rateReg}(object, newdata, groupName, groupLevels, level = 0.95,
  na.action, control = list(), ...)
}
\arguments{
\item{object}{An object used to dispatch a method.}

\item{...}{Other arguments for future usage.}

\item{data}{A data frame, list or environment containing the variables in
the model.  If not found in data, the variables are taken from
\code{environment(formula)}, usually the environment from which the
function is called.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process.}

\item{na.action}{A function that indicates what should the procedure do if
the data contains \code{NA}s.  The default is set by the na.action
setting of \code{\link[base]{options}}.  The "factory-fresh" default is
\code{\link[stats]{na.omit}}.  Other possible values inlcude
\code{\link[stats]{na.fail}}, \code{\link[stats]{na.exclude}}, and
\code{\link[stats]{na.pass}}.  \code{help(na.fail)} for details.}

\item{variance}{A character specifying the method for variance estimates.
The available options are \code{"LawlessNadeau"} (default) for Lawless
and Nadeau (1995) method, \code{"Poisson"} for Poisson process method,
and \code{"bootstrap"} for bootstrapping method.  Partial matching on
the names is allowed.}

\item{logConfInt}{A logical value. If \code{FALSE} (default), the confidence
interval are constructed based on the normality of the MCF
estimates. Otherwise, the confidence intervals of given level are
constucted based on the normality of logarithm of the MCF estimates.}

\item{level}{An optional numeric value indicating the confidence level
required. The default value is 0.95.}

\item{control}{An optional list to specify the time grid where the MCF is
estimated for \code{rateReg-class} object and other options for the
formula method.  The available named elements include
\itemize{
    \item \code{grid}: The time grid where MCF is estimated. A dense
        grid is suggested for further using the plot method.
    \item \code{length.out}: The length of grid points. The dafault
        value is 1,000.
    \item \code{from}: The starting point of grid. The default value is
        the left boundary knots (for \code{rateReg-class} object).
    \item \code{to}: The endpoint of grid. The default value is the
        right boundary knots (for \code{rateReg-class} object).
    \item \code{B}: The number of bootstrap replicates for using
        bootstrap method for variance estimates of sample MCF estimates.
        The default value is 200.
    \item \code{se.method}: The method used for SE estimates for
        bootstrap. The default method is \code{"sample.se"}, which takes
        the sample SE of point estimates from bootstrap samples.
    \item \code{ci.method}: The method used for confidence interval for
        bootstrap. The default method is the normal method based on
        normality.
}
The option \code{length.out}, \code{from}, \code{to} will be ignored if
\code{grid} is specified directly. Otherwise, the grid will be generated
by function \code{\link[base]{seq.int}} with specified \code{from},
\code{to} and \code{length.out}.}

\item{newdata}{An optional data frame. If specified, the data frame should
have the same column names as the covariate names appearing in the formula
of original fitting.}

\item{groupName}{An optional length-one charactor vector to specify the
name for grouping each unique row in \code{newdata}, such as "gender"
for "male" and "female". The default value is "group".}

\item{groupLevels}{An optional charactor vector to specify the levels for
each unique row in \code{newdata}, such as "treatment" and "control".
The default values are capital letters starting from "A".}
}
\value{
A \code{mcf.formula} or \code{mcf.rateReg} object.
The slots of a \code{mcf.formula} object include
\itemize{
    \item \code{level}: Confidence level specified.
    \item \code{MCF}: Mean cumulative function at each time point.
    \item \code{multiGroup}: A logical value indicating whether MCF
        is estimated for different specified group.
    \item \code{newdata}: Given dataset used to estimate MCF.
}
For the meaning of other slots, see \code{\link{rateReg}}.
}
\description{
An S4 class generic function that returns mean cumulative function (MCF)
estimates from a fitted model or computing the nonparametric MCF estimates
(also called the Nelson-Aalen estimator) from the sample data.
}
\details{
For \code{formula} object with \code{\link{Survr}} object as response, the
covariate specified at the right hand side of the formula should be either
\code{1} or any "linear" conbination of categorical variable in the data.
The former computes the overall sample MCF.  The latter computes the sample
MCF for each level of the combination of the categorical variable(s)
specified, respectively.  The sample MCF is also called Nelson-Aalen
nonparametric estimator (Nelson, 2003) and computed on each time point from
sample data.  The point estimate of sample MCF at each time point does not
assume any particular underlying model. The variance estimates at each time
point is computed following the Lawless and Nadeau method (LawLess and
Nadeau, 1995), the Poisson process method, or the bootstrap methods.  The
approximate confidence intervals are provided as well, which are constructed
based on the asymptotic normality of the MCF itself (by default) or the
logarithm of MCF.

For \code{\link{rateReg-class}} object, \code{mcf} estimates the baseline
MCF and its confidence interval at each time grid if argument \code{newdata}
is not specified.  Otherwise, \code{mcf} estimates MCF and its confidence
interval for the given newdata based on Delta-method.
}
\section{Methods (by class)}{
\itemize{
\item \code{formula}: Sample MCF from data.

\item \code{rateReg}: Estimated MCF from a fitted model.
}}

\examples{
library(reda)

### sample MCF
## Example 1. valve-seat data
valveMcf <- mcf(Survr(ID, Days, No.) ~ 1, data = valveSeats)
plot(valveMcf, conf.int = TRUE, mark.time = TRUE) + ggplot2::xlab("Days")

## Example 2. sample simulated data
simuMcf <- mcf(Survr(ID, time, event) ~ group + gender,
               data = simuDat, ID \%in\% 1 : 50, logConfInt = FALSE)
plot(simuMcf, conf.int = TRUE, lty = 1 : 4,
     legendName = "Treatment & Gender")

### estimate MCF difference between two groups
## one sample MCF object of two groups
mcf0 <- mcf(Survr(ID, time, event) ~ group, data = simuDat)
## two-sample pseudo-score tests
mcfDiff.test(mcf0)
## difference estimates over time
mcf0_diff <- mcfDiff(mcf0, testVariance = "none")
plot(mcf0_diff)

## or explicitly ask for the difference of two sample MCF
mcf1 <- mcf(Survr(ID, time, event) ~ 1, data = simuDat,
            subset = group \%in\% "Treat")
mcf2 <- mcf(Survr(ID, time, event) ~ 1, data = simuDat,
            subset = group \%in\% "Contr")
## perform two-sample tests and estimate difference at the same time
(mcf12_diff <- mcfDiff(mcf1, mcf2, level = 0.9))
plot(mcf12_diff)

### For estimated MCF from a fitted model,
### see examples given in function rateReg.
}
\references{
Lawless, J. F. and Nadeau, C. (1995). Some Simple Robust Methods for the
Analysis of Recurrent Events. \emph{Technometrics}, 37, 158--168.

Nelson, W. B. (2003). \emph{Recurrent Events Data Analysis for Product
Repairs, Disease Recurrences, and Other Applications} (Vol. 10). SIAM.
}
\seealso{
\code{\link{rateReg}} for model fitting;
\code{\link{mcfDiff}} for comparing two sample MCFs.
\code{\link{plot-method}} for plotting MCF.
}
