% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reda.R
\docType{package}
\name{reda-package}
\alias{reda-package}
\title{Recurrent Event Data Analysis}
\description{
The package \pkg{reda} mainly provides function \code{\link{rateReg}}
to fit parametric gamma frailty model with spline or piecewise constant
baseline rate function.
Another main function \code{\link{mcf}} computes and plots
the parametric mean cumulative function (MCF) from a fitted model
as well as the nonparametric sample MCF (Nelson-Aelson estimator)
for recurrent event data.
}
\details{
See vignettes for introduction and demonstration.
}

