% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show-method}
\alias{show-method}
\alias{show,Recur-method}
\alias{show,rateReg-method}
\alias{show,summary.rateReg-method}
\alias{show,summaryRateReg-method}
\alias{show,summary.Recur-method}
\alias{show,mcf.formula-method}
\alias{show,mcf.rateReg-method}
\alias{show,simEvent-method}
\alias{show,mcfDiff-method}
\alias{show,mcfDiff.test-method}
\title{Show an object.}
\usage{
\S4method{show}{Recur}(object)

\S4method{show}{rateReg}(object)

\S4method{show}{summary.rateReg}(object)

\S4method{show}{summary.Recur}(object)

\S4method{show}{mcf.formula}(object)

\S4method{show}{mcf.rateReg}(object)

\S4method{show}{simEvent}(object)

\S4method{show}{mcfDiff}(object)

\S4method{show}{mcfDiff.test}(object)
}
\arguments{
\item{object}{An object used to dispatch a method.}
}
\description{
S4 class methods that display objects produced from this package (similar to
S3 class \code{print} methods).
}
