% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{summaryRateReg-class}
\alias{summaryRateReg-class}
\title{An S4 Class to Represent Summary of a Fitted Model}
\description{
\code{summaryRateReg-class} is an S4 class with selective slots
of \code{rateReg-class} object.  See ``Slots'' for details.
\code{\link{summary,rateReg-method}} produces objects of this class.
}
\section{Slots}{

\describe{
\item{\code{call}}{Function call.}

\item{\code{spline}}{A character.}

\item{\code{knots}}{A numeric vector.}

\item{\code{Boundary.knots}}{A numeric vector.}

\item{\code{covarCoef}}{A numeric matrix.}

\item{\code{frailtyPar}}{A numeric matrix.}

\item{\code{degree}}{A nonnegative integer.}

\item{\code{baseRateCoef}}{A numeric matrix.}

\item{\code{logL}}{A numeric value.}
}}
\seealso{
\code{\link{summary,rateReg-method}} for details of slots.
}

