% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{Survr-class}
\alias{Survr-class}
\title{An S4 Class Representing Formula Response}
\description{
\code{Survr-class} is an S4 that represents a formula response for recurrent
event data model.  \code{\link{Survr}} produces objects of this class.
See ``Slots'' for details.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A numeric matrix object.}

\item{\code{ID}}{Charactrer vector for original subject identificator.}

\item{\code{check}}{Logical value indicating whether to performance data checking.}

\item{\code{ord}}{integer vector for increasingly ordering data by ID, time, and 1 -
event.}
}}

