% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{Survr-class}
\alias{Survr-class}
\title{An S4 Class Representing Formula Response}
\description{
The class \code{Survr} is an S4 that represents a formula response for
recurrent event data model.  The function \code{\link{Survr}} produces
objects of this class.  See ``Slots'' for details.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A numeric matrix object.}

\item{\code{ID}}{A charactrer vector for original subject identificator.}

\item{\code{check}}{A logical value indicating whether to performance data checking.}

\item{\code{ord}}{An integer vector for increasingly ordering data by \code{ID},
\code{time}, and \code{1 - event}.}
}}

\seealso{
\code{\link{Survr}}
}
